docs(documentation): add docs for ag-grid stand
consta-design-system/uikit#2568
