--- src/adapters/agGridAdapter/__stories__/AgGridAdapter.stories.css
@@ -1,8 +1,0 @@
-.AgGridAdapterStories {
-  width: 100%;
-  height: 100vh;
-
-  & .AgGridAdapter {
-    height: 100vh;
-  }
-}
