+++ src/components/DateInput/__stand__/DateInput.stand.mdx
@@ -1,0 +1,17 @@
+import { RemadeInfo } from '../../../stand/components/RemadeInfo';
+import { DateInputExample } from './examples/DateInputExample/DateInputExample';
+
+# DateInput
+
+<RemadeInfo
+  commponentName="DateInput"
+  apiUrl="https://www.ag-grid.com/react-data-grid/component-date/"
+/>
+
+```tsx
+import { DateInput } from '@consta/ag-grid-adapter/DateInput';
+```
+
+Компонент для выбора даты, используется в фильтрах — там, где пользователь может указать дату.
+
+<DateInputExample />
