+++ src/components/DateInput/__stand__/DateInput.stand.tsx
@@ -1,0 +1,10 @@
+import { createStand } from '##/stand/standConfig';
+
+export default createStand({
+  title: 'DateInput',
+  id: 'DateInput',
+  group: 'components',
+  description: 'Компонент для выбора даты.',
+  version: '0.1.1',
+  status: 'stable',
+});
