--- src/components/DateInput/__stories__/DateInput.stories.mdx
@@ -1,90 +1,0 @@
-import { Meta } from '@storybook/addon-docs/blocks';
-import {
-  RemadeInfo,
-  RemadeTableInfo,
-} from '@/__private__/storybook/components/RemadeInfo/RemadeInfo';
-import { DateInputExample } from './examples/DateInputExample/DateInputExample';
-
-<Meta title="Компоненты/DateInput" id="DateInput" />
-
-# DateInput
-
-<RemadeInfo
-  commponentName="DateInput"
-  apiUrl="https://www.ag-grid.com/react-data-grid/component-date/"
-/>
-
-```tsx
-import { DateInput } from '@consta/ag-grid-adapter/DateInput';
-```
-
-Компонент для выбора даты, используется в фильтрах — там, где пользователь может указать дату.
-
-## Свойства
-
-<RemadeTableInfo apiUrl="https://www.ag-grid.com/react-data-grid/component-date/" />
-
-| Параметры | Тип        | По умолчанию | Описание |
-| --------- | ---------- | ------------ | -------- |
-| `size?`   | `'s', 'm'` | `m`          | Размер   |
-
-## Пример
-
-```tsx
-const DateInputExample = () => {
-  const gridStyle = { height: '300px', width: '100%' };
-  const [rowData, setRowData] = useState();
-  const columnDefs = [
-    { field: 'athlete' },
-    { field: 'age', filter: 'agNumberColumnFilter' },
-    { field: 'country' },
-    { field: 'year' },
-    {
-      field: 'date',
-      minWidth: 190,
-      filter: 'agDateColumnFilter',
-    },
-    { field: 'sport' },
-    { field: 'gold', filter: 'agNumberColumnFilter' },
-    { field: 'silver', filter: 'agNumberColumnFilter' },
-    { field: 'bronze', filter: 'agNumberColumnFilter' },
-    { field: 'total', filter: false },
-  ];
-
-  const defaultColDef = {
-    editable: true,
-    sortable: true,
-    flex: 1,
-    minWidth: 100,
-    filter: true,
-    floatingFilter: true,
-    resizable: true,
-  };
-
-  const styleOptions = agGridAdapter();
-
-  useEffect(() => {
-    fetch('https://www.ag-grid.com/example-assets/olympic-winners.json')
-      .then((resp) => resp.json())
-      .then((data) => {
-        setRowData(data);
-      });
-  }, []);
-
-  return (
-    <div style={gridStyle} className="ag-theme-alpine">
-      <AgGridReact
-        {...styleOptions}
-        rowData={rowData}
-        columnDefs={columnDefs}
-        defaultColDef={defaultColDef}
-        components={{
-          agDateInput: DateInput,
-        }}
-      />
-    </div>
-  );
-};
-```
-
-<DateInputExample />
