+++ src/components/HeaderCell/__stand__/HeaderCell.stand.mdx
@@ -1,0 +1,19 @@
+import { RemadeInfo } from '../../../stand/components/RemadeInfo';
+import { HeaderCellExample } from './examples/HeaderCellExample/HeaderCellExample';
+
+# HeaderCell
+
+<RemadeInfo
+  commponentName="HeaderCell"
+  apiUrl="https://www.ag-grid.com/react-data-grid/component-header/"
+/>
+
+```tsx
+import { HeaderCell } from '@consta/ag-grid-adapter/HeaderCell';
+```
+
+С помощью этого компонента можно отдельно от всей таблицы управлять размерами и выравниванием текста в заголовке, а также указывать необходимые свойства компонента
+[Text](https://uikit.consta.design/?path=/story/components-text--playground).
+Применяется к заголовку определённого столбца.
+
+<HeaderCellExample />
