+++ src/components/HeaderCell/__stand__/HeaderCell.stand.tsx
@@ -1,0 +1,10 @@
+import { createStand } from '##/stand/standConfig';
+
+export default createStand({
+  title: 'HeaderCell',
+  id: 'HeaderCell',
+  group: 'components',
+  description: 'Элемент заголовка таблицы',
+  version: '0.1.1',
+  status: 'stable',
+});
