+++ src/components/LoadingCell/__stand__/LoadingCell.stand.mdx
@@ -1,0 +1,19 @@
+import { RemadeInfo } from '../../../stand/components/RemadeInfo';
+import { LoadingCellExample } from './examples/LoadingCellExample/LoadingCellExample';
+
+# LoadingCell
+
+<RemadeInfo
+  commponentName="LoadingCell"
+  apiUrl="https://www.ag-grid.com/react-data-grid/component-loading-cell-renderer/"
+/>
+
+```tsx
+import { LoadingCell } from '@consta/ag-grid-adapter/LoadingCell';
+```
+
+Компонент `LoadingCell` формирует дополнительную строку с сообщением. Эта строка появляется при прокрутке таблицы и видна, пока загружаются следующие строчки.
+
+Зачем это нужно: представьте, что пользователь так быстро прокручивает таблицу до конца, что строки не успевают загружаться. В этот момент можно подсказать пользователю, что это ещё не конец: нужно немного подождать, и информация загрузится.
+
+<LoadingCellExample />
