+++ src/components/LoadingCell/__stand__/LoadingCell.stand.tsx
@@ -1,0 +1,10 @@
+import { createStand } from '##/stand/standConfig';
+
+export default createStand({
+  title: 'LoadingCell',
+  id: 'LoadingCell',
+  group: 'components',
+  description: 'Строка загрузки данных.',
+  version: '0.1.1',
+  status: 'stable',
+});
