+++ src/components/Overlay/__stand__/Overlay.stand.mdx
@@ -1,0 +1,17 @@
+import { RemadeInfo } from '../../../stand/components/RemadeInfo';
+import { OverlayExample } from './examples/OverlayExample/OverlayExample';
+
+# Overlay
+
+<RemadeInfo
+  commponentName="Overlay"
+  apiUrl="https://www.ag-grid.com/react-data-grid/component-overlay/"
+/>
+
+```tsx
+import { Overlay } from '@consta/ag-grid-adapter/Overlay';
+```
+
+Этот компонент создаёт полупрозрачный слой поверх таблицы (наложение) и показывает сверху сообщение или анимированный загрузчик.
+
+<OverlayExample />
