+++ src/components/Overlay/__stand__/Overlay.stand.tsx
@@ -1,0 +1,10 @@
+import { createStand } from '##/stand/standConfig';
+
+export default createStand({
+  title: 'Overlay',
+  id: 'Overlay',
+  group: 'components',
+  description: 'Наложение поверх таблицы для сообщения или лоудера.',
+  version: '0.1.1',
+  status: 'stable',
+});
