--- src/components/Overlay/__stories__/Overlay.stories.mdx
@@ -1,129 +1,0 @@
-import { Meta } from '@storybook/addon-docs/blocks';
-import { OverlayExample } from './examples/OverlayExample/OverlayExample';
-import {
-  RemadeInfo,
-  RemadeTableInfo,
-} from '@/__private__/storybook/components/RemadeInfo/RemadeInfo';
-
-<Meta title="Компоненты/Overlay" id="Overlay" />
-
-# Overlay
-
-<RemadeInfo
-  commponentName="Overlay"
-  apiUrl="https://www.ag-grid.com/react-data-grid/component-overlay/"
-/>
-
-```tsx
-import { Overlay } from '@consta/ag-grid-adapter/Overlay';
-```
-
-Этот компонент создаёт полупрозрачный слой поверх таблицы (наложение) и показывает сверху сообщение или анимированный загрузчик.
-
-## Свойства
-
-<RemadeInfoTable apiUrl="https://www.ag-grid.com/react-data-grid/component-overlay/" />
-
-| Параметры  | Тип                    | По умолчанию            | Описание                                                                          |
-| ---------- | ---------------------- | ----------------------- | --------------------------------------------------------------------------------- |
-| `size?`    | `'s', 'm'`             | `'m'`                   | Размер                                                                            |
-| `type?`    | `'loading', 'message'` | `'loading'`             | Что показывать поверх наложения: `'loading'` — загрузчик, `'message'` — сообщение |
-| `message?` | `string`               | `Результаты не найдены` | Текст сообщения                                                                   |
-| `icon?`    | `React.FC<IconProps>`  | -                       | Иконка для загрузчика                                                             |
-
-## Пример
-
-```tsx
-const OverlayExample = () => {
-  const gridRef = useRef<AgGridReact>(null);
-  const containerStyle = useMemo(
-    () => ({ width: '100%', height: '300px' }),
-    [],
-  );
-  const gridStyle = useMemo(() => ({ height: '100%', width: '100%' }), []);
-  const [rowData, setRowData] = useState();
-  const columnDefs = [
-    { field: 'athlete', width: 150 },
-    { field: 'age', width: 90 },
-    { field: 'country', width: 120 },
-    { field: 'year', width: 90 },
-    { field: 'date', width: 110 },
-    { field: 'sport', width: 110 },
-    { field: 'gold', width: 100 },
-    { field: 'silver', width: 100 },
-    { field: 'bronze', width: 100 },
-    { field: 'total', width: 100 },
-  ];
-
-  const defaultColDef = useMemo(() => {
-    return {
-      editable: true,
-      sortable: true,
-      flex: 1,
-      minWidth: 100,
-      filter: true,
-      resizable: true,
-    };
-  }, []);
-  const loadingOverlayComponent = useMemo(() => {
-    return (props: ILoadingOverlayParams) => (
-      <Overlay type="loading" {...props} />
-    );
-  }, []);
-
-  const noRowsOverlayComponent = useMemo(() => {
-    return (props: ILoadingOverlayParams) => (
-      <Overlay type="message" {...props} />
-    );
-  }, []);
-
-  const styleOptions = agGridAdapter();
-
-  const onGridReady = useCallback(() => {
-    fetch('https://www.ag-grid.com/example-assets/olympic-winners.json')
-      .then((resp) => resp.json())
-      .then((data) => {
-        setRowData(data);
-      });
-  }, []);
-
-  const onBtShowLoading = () => {
-    gridRef?.current?.api.showLoadingOverlay();
-  };
-
-  const onBtShowNoRows = () => {
-    gridRef?.current?.api.showNoRowsOverlay();
-  };
-
-  const onBtHide = () => {
-    gridRef?.current?.api.hideOverlay();
-  };
-
-  return (
-    <div style={containerStyle} className={cnOverlayExample()}>
-      <div className={cnOverlayExample('Container')}>
-        <div className={cnOverlayExample('Buttons')}>
-          <Button onClick={onBtShowLoading} label="Показать загрузчик" />
-          <Button onClick={onBtShowNoRows} label="Показать сообщение" />
-          <Button onClick={onBtHide} label="Закрыть все" />
-        </div>
-
-        <div style={gridStyle}>
-          <AgGridReact
-            ref={gridRef}
-            rowData={rowData}
-            {...styleOptions}
-            columnDefs={columnDefs}
-            defaultColDef={defaultColDef}
-            loadingOverlayComponent={loadingOverlayComponent}
-            noRowsOverlayComponent={noRowsOverlayComponent}
-            onGridReady={onGridReady}
-          />
-        </div>
-      </div>
-    </div>
-  );
-};
-```
-
-<OverlayExample />
