+++ src/components/TableCell/__stand__/TableCell.stand.mdx
@@ -1,0 +1,18 @@
+import { RemadeInfo } from '../../../stand/components/RemadeInfo';
+import { TableCellExample } from './examples/TableCellExample/TableCellExample';
+
+# TableCell
+
+<RemadeInfo
+  commponentName="TableCell"
+  apiUrl="https://www.ag-grid.com/react-data-grid/component-cell-renderer/"
+/>
+
+```tsx
+import { TableCell } from '@consta/ag-grid-adapter/TableCell';
+```
+
+Этот компонент кастомизирует ячейку таблицы. С его помощью можно поменять размер текста, выравнивание, добавить в ячейку иконку из [библиотеки иконок](https://uikit.consta.design/?path=/story/components-icons--playground), а также указывать необходимые свойства компонента
+[Text](https://uikit.consta.design/?path=/story/components-text--playground).
+
+<TableCellExample />
