+++ src/components/TableCell/__stand__/TableCell.stand.tsx
@@ -1,0 +1,10 @@
+import { createStand } from '##/stand/standConfig';
+
+export default createStand({
+  title: 'TableCell',
+  id: 'TableCell',
+  group: 'components',
+  description: 'Этот компонент кастомизирует ячейку таблицы.',
+  version: '0.1.1',
+  status: 'stable',
+});
