--- src/components/TableCell/__stories__/TableCell.stories.mdx
@@ -1,91 +1,0 @@
-import { Meta } from '@storybook/addon-docs/blocks';
-import {
-  RemadeInfo,
-  RemadeTableInfo,
-} from '@/__private__/storybook/components/RemadeInfo/RemadeInfo';
-import { TableCellExample } from './examples/TableCellExample/TableCellExample';
-
-<Meta title="Компоненты/TableCell" id="TableCell" />
-
-# TableCell
-
-<RemadeInfo
-  commponentName="TableCell"
-  apiUrl="https://www.ag-grid.com/react-data-grid/component-cell-renderer/"
-/>
-
-```tsx
-import { TableCell } from '@consta/ag-grid-adapter/TableCell';
-```
-
-Этот компонент кастомизирует ячейку таблицы. С его помощью можно поменять размер текста, выравнивание, добавить в ячейку иконку из [библиотеки иконок](https://uikit.consta.design/?path=/story/components-icons--playground), а также указывать необходимые свойства компонента
-[Text](https://uikit.consta.design/?path=/story/components-text--playground).
-
-## Свойства
-
-<RemadeTableInfo apiUrl="https://www.ag-grid.com/react-data-grid/component-cell-renderer/" />
-
-| Параметры   | Тип                         | По умолчанию | Описание                       |
-| ----------- | --------------------------- | ------------ | ------------------------------ |
-| `size?`     | `'s', 'm'`                  | `'m'`        | Размер                         |
-| `align?`    | `'left', 'right', 'center'` | `'left'`     | Выравнивание контента в ячейке |
-| `children?` | `React.ReactElement`        | -            | Дополнительный контент         |
-| `icon?`     | `React.FC<IconProps>`       | -            | Иконка                         |
-
-## Пример
-
-```tsx
-export const TableCellExample = () => {
-  const gridStyle = { height: '300px', width: '100%' };
-  const [rowData, setRowData] = useState();
-  const columnDefs = [
-    { field: 'athlete' },
-    { field: 'year' },
-    {
-      field: 'gold',
-      cellRenderer: (props: ICellRendererParams) => (
-        <TableCell icon={IconCrown} {...props} />
-      ),
-    },
-    { field: 'silver', cellRenderer: TableCell },
-    { field: 'bronze', cellRenderer: TableCell },
-    { field: 'total', minWidth: 175, cellRenderer: TableCell },
-  ];
-
-  const defaultColDef = {
-    editable: true,
-    sortable: true,
-    flex: 1,
-    minWidth: 100,
-    filter: true,
-    resizable: true,
-    cellRendererParams: {
-      view: 'warning',
-      weight: 'bold',
-    },
-  };
-
-  const styleOptions = agGridAdapter();
-
-  useEffect(() => {
-    fetch('https://www.ag-grid.com/example-assets/olympic-winners.json')
-      .then((resp) => resp.json())
-      .then((data) => {
-        setRowData(data);
-      });
-  }, []);
-
-  return (
-    <div style={gridStyle}>
-      <AgGridReact
-        {...styleOptions}
-        rowData={rowData}
-        columnDefs={columnDefs}
-        defaultColDef={defaultColDef}
-      />
-    </div>
-  );
-};
-```
-
-<TableCellExample />
