+++ src/components/TableEditorNumber/__stand__/TableEditorNumber.stand.mdx
@@ -1,0 +1,18 @@
+import { RemadeInfo } from '../../../stand/components/RemadeInfo';
+import { TableEditorNumberExample } from './examples/TableEditorNumberExample/TableEditorNumberExample';
+
+# TableEditorNumber
+
+<RemadeInfo
+  commponentName="TableEditorNumber"
+  apiUrl="https://www.ag-grid.com/react-data-grid/component-cell-editor/"
+/>
+
+```tsx
+import { TableEditorNumber } from '@consta/ag-grid-adapter/TableEditorNumber';
+```
+
+Компонент редактора для изменения численных значений в ячейке таблицы. Позволяет менять размер редактора без использования адаптера, а также указывать необходимые свойства компонента
+[TextField](https://uikit.consta.design/?path=/story/components-textfield--playground).
+
+<TableEditorNumberExample />
