+++ src/components/TableEditorNumber/__stand__/TableEditorNumber.stand.tsx
@@ -1,0 +1,11 @@
+import { createStand } from '##/stand/standConfig';
+
+export default createStand({
+  title: 'TableEditorNumber',
+  id: 'TableEditorNumber',
+  group: 'components',
+  description:
+    'Компонент редактора для изменения числовых значений в ячейке таблицы.',
+  version: '0.1.1',
+  status: 'stable',
+});
