+++ src/components/TableEditorText/__stand__/TableEditorText.dev.stand.mdx
@@ -1,0 +1,87 @@
+import { MdxMenu } from '@consta/stand';
+import { RemadeTableInfo } from '../../../stand/components/RemadeInfo';
+import { TableEditorTextExample } from './examples/TableEditorTextExample/TableEditorTextExample';
+
+<MdxMenu>
+
+- [Свойства](#свойства)
+- [Пример](#пример)
+
+</MdxMenu>
+
+Компонент редактора для изменения строковых значений в ячейке таблицы. Позволяет менять размер редактора без использования адаптера, а также указывать необходимые свойства компонента
+[TextField](https://uikit.consta.design/?path=/story/components-textfield--playground).
+
+## Свойства
+
+<RemadeTableInfo apiUrl="https://www.ag-grid.com/react-data-grid/component-cell-editor/" />
+
+| Параметры | Тип        | По умолчанию | Описание |
+| --------- | ---------- | ------------ | -------- |
+| `size?`   | `'s', 'm'` | `m`          | Размер   |
+
+## Пример
+
+```tsx
+const TableEditorTextExample = () => {
+  const rowData = [
+    { name: 'Bob', mood: 'Happy', number: 10 },
+    { name: 'Harry', mood: 'Sad', number: 3 },
+    { name: 'Sally', mood: 'Happy', number: 20 },
+    { name: 'Mary', mood: 'Sad', number: 5 },
+    { name: 'John', mood: 'Happy', number: 15 },
+    { name: 'Jack', mood: 'Happy', number: 25 },
+    { name: 'Sue', mood: 'Sad', number: 43 },
+    { name: 'Sean', mood: 'Sad', number: 1335 },
+    { name: 'Niall', mood: 'Happy', number: 2 },
+    { name: 'Alberto', mood: 'Happy', number: 123 },
+    { name: 'Fred', mood: 'Sad', number: 532 },
+    { name: 'Jenny', mood: 'Happy', number: 34 },
+    { name: 'Larry', mood: 'Happy', number: 13 },
+  ]
+
+  const columnDefs = [
+    {
+      headerName: 'Doubling',
+      field: 'name',
+      cellEditor: TableEditorText,
+      cellEditorPopup: true,
+      editable: true,
+      width: 300,
+    },
+    {
+      field: 'mood',
+      width: 300,
+    },
+    {
+      headerName: 'Numeric',
+      field: 'number',
+      width: 280,
+    },
+  ]
+
+  const styleOptions = agGridAdapter()
+
+  return (
+    <div style={{ width: '100%', height: '300px' }}>
+      <AgGridReact
+        {...styleOptions}
+        columnDefs={columnDefs}
+        rowData={rowData}
+        defaultColDef={{
+          cellEditorParams: {
+            leftSide: IconUser,
+          },
+          editable: true,
+          sortable: true,
+          flex: 1,
+          minWidth: 100,
+          filter: true,
+          resizable: true,
+        }}
+      />
+    </div>
+  )
+```
+
+<TableEditorTextExample />
