+++ src/components/TableEditorText/__stand__/TableEditorText.stand.mdx
@@ -1,0 +1,18 @@
+import { RemadeInfo } from '../../../stand/components/RemadeInfo';
+import { TableEditorTextExample } from './examples/TableEditorTextExample/TableEditorTextExample';
+
+# TableEditorText
+
+<RemadeInfo
+  commponentName="TableEditorText"
+  apiUrl="https://www.ag-grid.com/react-data-grid/component-cell-editor/"
+/>
+
+```tsx
+import { TableEditorNumber } from '@consta/ag-grid-adapter/TableEditorText';
+```
+
+Компонент редактора для изменения строковых значений в ячейке таблицы. Позволяет менять размер редактора без использования адаптера, а также указывать необходимые свойства компонента
+[TextField](https://uikit.consta.design/?path=/story/components-textfield--playground).
+
+<TableEditorTextExample />
