+++ src/components/TableEditorText/__stand__/TableEditorText.stand.tsx
@@ -1,0 +1,11 @@
+import { createStand } from '##/stand/standConfig';
+
+export default createStand({
+  title: 'TableEditorText',
+  id: 'TableEditorText',
+  group: 'components',
+  description:
+    'Компонент редактора для изменения строковых значений в ячейке таблицы.',
+  version: '0.1.1',
+  status: 'stable',
+});
