+++ src/components/Tooltip/__stand__/Tooltip.dev.stand.mdx
@@ -1,0 +1,87 @@
+import { MdxMenu } from '@consta/stand';
+import { RemadeTableInfo } from '../../../stand/components/RemadeInfo';
+import { TooltipExample } from './examples/TooltipExample/TooltipExample';
+
+<MdxMenu>
+
+- [Свойства](#свойства)
+- [Пример](#пример)
+
+</MdxMenu>
+
+Обёртка для тултипа. Тултип привязывается к колонке и всплывает при наведении на неё. В тултипе можно показать любую дополнительную информацию.
+
+## Свойства
+
+<RemadeTableInfo apiUrl="https://www.ag-grid.com/react-data-grid/component-tooltip/" />
+
+| Параметры   | Тип                                            | По умолчанию | Описание        |
+| ----------- | ---------------------------------------------- | ------------ | --------------- |
+| `children?` | `(data: ITooltipParams) => React.ReactElement` | -            | Контент тултипа |
+
+## Пример
+
+```tsx
+const TooltipExample = () => {
+  const gridStyle = { height: '300px', width: '100%' };
+  const [rowData, setRowData] = useState();
+  const columnDefs = [
+    {
+      field: 'athlete',
+      minWidth: 150,
+      tooltipField: 'athlete',
+    },
+    { field: 'age' },
+    { field: 'country', minWidth: 130, tooltipField: 'country' },
+    { field: 'year' },
+    { field: 'date' },
+    { field: 'sport' },
+    { field: 'gold' },
+    { field: 'silver' },
+    { field: 'bronze' },
+    { field: 'total' },
+  ];
+
+  const defaultColDef = {
+    editable: true,
+    sortable: true,
+    flex: 1,
+    minWidth: 100,
+    filter: true,
+    resizable: true,
+    tooltipComponent: Tooltip,
+    tooltipComponentParams: {
+      children: (props: ITooltipParams) => (
+        <Text>
+          Имя: {props.data.athlete}, возраст: {props.data.age}
+        </Text>
+      ),
+    },
+  };
+
+  const styleOptions = agGridAdapter();
+
+  useEffect(() => {
+    fetch('https://www.ag-grid.com/example-assets/olympic-winners.json')
+      .then((resp) => resp.json())
+      .then((data) => {
+        setRowData(data);
+      });
+  }, []);
+
+  return (
+    <div style={gridStyle}>
+      <AgGridReact
+        {...styleOptions}
+        rowData={rowData}
+        columnDefs={columnDefs}
+        defaultColDef={defaultColDef}
+        tooltipShowDelay={0}
+        tooltipHideDelay={1000}
+      />
+    </div>
+  );
+};
+```
+
+<TooltipExample />
