+++ src/components/Tooltip/__stand__/Tooltip.stand.mdx
@@ -1,0 +1,17 @@
+import { RemadeInfo } from '../../../stand/components/RemadeInfo';
+import { TooltipExample } from './examples/TooltipExample/TooltipExample';
+
+# Tooltip
+
+<RemadeInfo
+  commponentName="Tooltip"
+  apiUrl="https://www.ag-grid.com/react-data-grid/component-tooltip/"
+/>
+
+```tsx
+import { Tooltip } from '@consta/ag-grid-adapter/Tooltip';
+```
+
+Обёртка для тултипа. Тултип привязывается к колонке и всплывает при наведении на неё. В тултипе можно показать любую дополнительную информацию.
+
+<TooltipExample />
