+++ src/components/Tooltip/__stand__/Tooltip.stand.tsx
@@ -1,0 +1,10 @@
+import { createStand } from '##/stand/standConfig';
+
+export default createStand({
+  title: 'Tooltip',
+  id: 'Tooltip',
+  group: 'components',
+  description: 'Обёртка для тултипа.',
+  version: '0.1.1',
+  status: 'stable',
+});
